Replace non-negative isize with usize
- FineCosine now indexes by usize; call sites drop their isize casts.
- The two R_DrawMasked drawseg loops iterate (0..ds_p).rev() instead of
  counting an isize down to -1.
- Drop the intptr_t alias; x11 uses ptr.add instead of ptr.offset.

isize stays where the value can be negative: ClipArray::Openings base and
column index, r_draw source/fuzz offsets, and the -3 column-source delta.

+++ rust-doomgeneric/engine/src/tables.rs
@@ -1167,11 +1167,11 @@ pub static finesine: [i32; 10240] = [
     65535, 65535, 65535, 65535, 65535, 65535, 65535,
 ];
 pub struct FineCosine;
-impl core::ops::Index<isize> for FineCosine {
+impl core::ops::Index<usize> for FineCosine {
     type Output = fixed_t;
     #[inline]
-    fn index(&self, index: isize) -> &fixed_t {
-        &finesine[(FINEANGLES / 4) as usize + index as usize]
+    fn index(&self, index: usize) -> &fixed_t {
+        &finesine[(FINEANGLES / 4) as usize + index]
     }
 }
 pub static finecosine: FineCosine = FineCosine;
